Change Wrap Algorithm Completely
Before this commit, we created a Docker image in tarball format locally
and uploaded it to Docker (via `load`), effectively having full control
over the image creation process.

Unfortunately, due to the new format that has been developed by Docker
we are left without documentation, examples or even APIs that make this
approach feasible if support is neccessary for version 1.10 (which I
think is crucial). In my opininion, Docker really failed to provide a
contigency path for API users like us.

* If a base image is set (such as busybox) we create a container, put
our files into it and commit it (with config) into a new repository.
This works fine, but (probably) involves copying the files twice instead
of once.

* If a base image is not set (similar to "FROM scratch"), we can't
create a container, since we don't have an image (scratch is not a
createable image). So, to bootstrap the image, we `import` a tar archive
into Docker, but (and this is again API-designed-at-Docker^tm): Docker
import does not allow setting an image configuration (at least the
_documented_ API. The CLI can do this, and it uses undocumented
parameters for this.) So instead, we import an image, create a container
out of it, commit it to the actual repository we want with the config we
want, and drop the container again.